test: ColumnServiceTest
- 칼럼 추가
- 칼럼에 카드 추가
- 칼럼을 가져오기
- 모든 칼럼을 가져오기
- 칼럼에서 카드를 삭제하기

+++ backend/src/main/resources/schema.sql
@@ -1,41 +1,41 @@
 CREATE TABLE ToDoList (
-    id   int AUTO_INCREMENT PRIMARY KEY,
+    id   BIGINT AUTO_INCREMENT PRIMARY KEY,
     name varchar(50) NOT NULL
 );
 
 CREATE TABLE ToDoList_Column (
-    ToDoList int,
-    Column int,
-    ToDoList_key int
+    ToDoList BIGINT,
+    Column BIGINT,
+    ToDoList_key BIGINT
 );
 
 CREATE TABLE Column (
-    id   int AUTO_INCREMENT PRIMARY KEY,
+    id   BIGINT AUTO_INCREMENT PRIMARY KEY,
     name varchar(50) NOT NULL
 );
 
 CREATE TABLE Column_Card (
-    Column int,
-    Card int,
-    Column_key int
+    Column BIGINT,
+    Card BIGINT,
+    Column_key BIGINT
 );
 
 CREATE TABLE Card (
-    id    int AUTO_INCREMENT PRIMARY KEY,
+    id    BIGINT AUTO_INCREMENT PRIMARY KEY,
     title varchar(50),
     body  varchar(500)
 );
 
 CREATE TABLE user (
-    id       int AUTO_INCREMENT PRIMARY KEY,
+    id       BIGINT AUTO_INCREMENT PRIMARY KEY,
     userid   VARCHAR(45) NOT NULL UNIQUE,
     password varchar(45) NOT NULL,
     name     varchar(45) NOT NULL
 );
 
 CREATE TABLE activity (
-    id          int AUTO_INCREMENT PRIMARY KEY,
-    action_user int,
+    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
+    action_user BIGINT,
     action      char(10),
     title       varchar(50),
     from_column varchar(50),
